+++ Model/Config/Source/Log.php
@@ -13,7 +13,7 @@ class Log implements \Magento\Framework\Option\ArrayInterface
     {
         return [
             ['value' => 0, 'label' => __('Disabled')],
-            ['value' => 1, 'label' => __('Only Reepay API')],
+            ['value' => 1, 'label' => __('Only Billwerk+ API')],
             ['value' => 2, 'label' => __('Debug mode')],
         ];
     }